Move verbatim lyrics to isolate utils
