+++ cross-platform/MusicLyricApp/Core/Service/Music/NetEaseMusicApi.cs
@@ -144,7 +144,7 @@ public class NetEaseMusicApi(Func<string> cookieFunc) : MusicCacheableApi
         {
             if (resp.Yrc != null)
             {
-                vo.Lyric = LyricUtils.DealVerbatimLyric4NetEaseMusic(resp.Yrc.Lyric);
+                vo.Lyric = VerbatimLyricUtils.DealVerbatimLyric4NetEaseMusic(resp.Yrc.Lyric);
             }
             // not support translate && Transliteration in common mode
         }
