+++ cross-platform/MusicLyricApp/Core/Utils/LyricUtils.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
-using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using hyjiacan.py4n;
@@ -18,13 +17,8 @@ namespace MusicLyricApp.Core.Utils;
 /// </summary>
 public static partial class LyricUtils
 {
-    private static readonly Regex VerbatimRegex4QQMusic = GetVerbatimRegex4QQMusicRegex();
-
-    private static readonly Regex VerbatimRegex4NetEaseMusic = GetVerbatimRegex4NetEaseMusicRegex();
-
-    public static readonly Regex VerbatimLegalPrefixRegex = GetVerbatimLegalPrefixRegex();
-
-    public static readonly Regex CommonLegalPrefixRegex = GetCommonLegalPrefixRegex();
+    [GeneratedRegex(LyricLineVo.TimestampPattern)]
+    public static partial Regex GetCommonLegalPrefixRegex();
 
     private const PinyinFormat PinyinDefineFormat =
         PinyinFormat.WITH_TONE_MARK | PinyinFormat.LOWERCASE | PinyinFormat.WITH_U_UNICODE;
@@ -48,7 +42,7 @@ public static partial class LyricUtils
         {
             for (var i = 0; i < voListList.Count; i++)
             {
-                voListList[i] = FormatSubLineLyric(voListList[i], timestampFormat, dotType);
+                voListList[i] = VerbatimLyricUtils.FormatSubLineLyric(voListList[i], timestampFormat, dotType);
             }
         }
 
@@ -74,189 +68,6 @@ public static partial class LyricUtils
             };
 
             res.Add(line);
-        }
-
-        return res;
-    }
-
-    public static string ConvertVerbatimLyricToCommonLyric4QQMusic(string input)
-    {
-        if (string.IsNullOrWhiteSpace(input))
-        {
-            return string.Empty;
-        }
-
-        var defaultParam = new ConfigBean();
-        var sb = new StringBuilder();
-
-        foreach (var line in SplitLrc(input))
-        {
-            var matches = VerbatimLegalPrefixRegex.Matches(line);
-            if (matches.Count > 0)
-            {
-                var group = matches[0].Groups[0];
-                int leftParenthesesIndex = group.Index, parenthesesLength = group.Length;
-
-                // [70,80]
-                var timeStr = line.Substring(leftParenthesesIndex, parenthesesLength);
-                // 70
-                var timestamp = long.Parse(timeStr.Split(',')[0].Trim()[1..]);
-                var lyricTimestamp = new LyricTimestamp(timestamp);
-
-                var content = VerbatimRegex4QQMusic.Replace(line[parenthesesLength..], string.Empty);
-
-                sb.Append(lyricTimestamp.PrintTimestamp(defaultParam.LrcTimestampFormat, defaultParam.DotType))
-                    .Append(content);
-            }
-            else
-            {
-                sb.Append(line);
-            }
-
-            sb.Append(Environment.NewLine);
-        }
-
-        return sb.ToString();
-    }
-
-    public static string DealVerbatimLyric4QQMusic(string originLrc)
-    {
-        if (string.IsNullOrWhiteSpace(originLrc))
-        {
-            return string.Empty;
-        }
-
-        var defaultParam = new ConfigBean();
-        var sb = new StringBuilder();
-
-        foreach (var line in SplitLrc(originLrc))
-        {
-            // skip illegal verbatim line, eg: https://y.qq.com/n/ryqq/songDetail/000sNzbP2nHGs2
-            if (!line.EndsWith(")"))
-            {
-                continue;
-            }
-
-            var matches = VerbatimRegex4QQMusic.Matches(line);
-            if (matches.Count > 0)
-            {
-                int contentStartIndex = 0, i = 0;
-
-                do
-                {
-                    var curMatch = matches[i];
-                    var group = curMatch.Groups[0];
-                    int leftParenthesesIndex = group.Index, parenthesesLength = group.Length;
-
-                    // (404,202)
-                    var timeStr = line.Substring(leftParenthesesIndex, parenthesesLength);
-                    // 404
-                    var timestamp = long.Parse(timeStr.Split(',')[0].Trim()[1..]);
-                    var lyricTimestamp = new LyricTimestamp(timestamp);
-
-                    var content = line.Substring(contentStartIndex, leftParenthesesIndex - contentStartIndex);
-                    // 首次执行，去除全行时间戳
-                    if (i == 0)
-                    {
-                        content = new LyricLineVo(content).Content;
-                    }
-
-                    contentStartIndex = leftParenthesesIndex + parenthesesLength;
-
-                    sb.Append(lyricTimestamp.PrintTimestamp(defaultParam.LrcTimestampFormat, defaultParam.DotType))
-                        .Append(content);
-
-                    // 最后一次执行，增加行结束时间戳
-                    if (i == matches.Count - 1)
-                    {
-                        // 202
-                        var timeCostStr = timeStr.Split(',')[1].Trim();
-                        var timeCost = long.Parse(timeCostStr[..^1]);
-
-                        sb.Append(lyricTimestamp.Add(timeCost)
-                            .PrintTimestamp(defaultParam.LrcTimestampFormat, defaultParam.DotType));
-                    }
-                } while (++i < matches.Count);
-            }
-            else
-            {
-                sb.Append(line);
-            }
-
-            sb.Append(Environment.NewLine);
-        }
-
-        return sb.ToString();
-    }
-
-    public static string DealVerbatimLyric4NetEaseMusic(string originLrc)
-    {
-        if (string.IsNullOrWhiteSpace(originLrc))
-        {
-            return string.Empty;
-        }
-
-        var defaultParam = new ConfigBean();
-        var sb = new StringBuilder();
-
-        foreach (var originLine in SplitLrc(originLrc))
-        {
-            var line = originLine;
-
-            var prefixMatch = VerbatimLegalPrefixRegex.Match(line);
-            if (prefixMatch.Success)
-            {
-                var prefix = line.Substring(prefixMatch.Index, prefixMatch.Length);
-
-                // remove prefix
-                line = line[(prefixMatch.Index + prefixMatch.Length)..];
-
-                Match match;
-                while ((match = VerbatimRegex4NetEaseMusic.Match(line)).Success)
-                {
-                    var group = match.Groups[0];
-
-                    // (390,220,0)
-                    var timeStr = line.Substring(group.Index, group.Length);
-                    // 390
-                    var timestamp = long.Parse(timeStr.Split(',')[0].Trim()[1..]);
-                    var lyricTimestamp = new LyricTimestamp(timestamp);
-
-                    line = line.Replace(timeStr,
-                        lyricTimestamp.PrintTimestamp(defaultParam.LrcTimestampFormat, defaultParam.DotType));
-                }
-
-                // 追加行尾 timestamp
-                var prefixArr = prefix.Split(',');
-                var lastTimestamp =
-                    new LyricTimestamp(long.Parse(prefixArr[0].Trim()[1..]) + long.Parse(prefixArr[1].Trim()[..^1]));
-                line += lastTimestamp.PrintTimestamp(defaultParam.LrcTimestampFormat, defaultParam.DotType);
-            }
-
-            sb.Append(line);
-            sb.Append(Environment.NewLine);
-        }
-
-        return sb.ToString();
-    }
-
-    /// <summary>
-    /// need try split sub lyricLineVO, resolve verbatim lyric mode
-    /// </summary>
-    /// <returns></returns>
-    private static List<LyricLineVo> FormatSubLineLyric(List<LyricLineVo> vos, string timestampFormat,
-        DotTypeEnum dotType)
-    {
-        var res = new List<LyricLineVo>();
-        foreach (var vo in vos)
-        {
-            var sb = new StringBuilder();
-            foreach (var subVo in LyricLineVo.Split(vo))
-            {
-                sb.Append(subVo.Timestamp.PrintTimestamp(timestampFormat, dotType) + subVo.Content);
-            }
-
-            res.Add(new LyricLineVo(sb.ToString()));
         }
 
         return res;
@@ -346,7 +157,7 @@ public static partial class LyricUtils
         return res;
     }
 
-    private static string[] SplitLrc(string lrc)
+    public static string[] SplitLrc(string lrc)
     {
         return (lrc ?? "")
             .Replace("\r\n", "\n")
@@ -812,16 +623,4 @@ public static partial class LyricUtils
                 return LyricsTypeEnum.ORIGIN_TRANS;
         }
     }
-
-    [GeneratedRegex(LyricLineVo.TimestampPattern)]
-    private static partial Regex GetCommonLegalPrefixRegex();
-
-    [GeneratedRegex(@"\[\d+,\d+\]")]
-    private static partial Regex GetVerbatimLegalPrefixRegex();
-
-    [GeneratedRegex(@"\(\d+,\d+,\d+\)")]
-    private static partial Regex GetVerbatimRegex4NetEaseMusicRegex();
-
-    [GeneratedRegex(@"\(\d+,\d+\)")]
-    private static partial Regex GetVerbatimRegex4QQMusicRegex();
 }
